Removed `Texts` class, moved to using `safe_load_texts`
