+++ node_cli/cli/exit.py
@@ -24,10 +24,10 @@ import click
 from node_cli.utils.print_formatters import print_exit_status
 from node_cli.utils.helper import error_exit, get_request, post_request, abort_if_false
 from node_cli.utils.exit_codes import CLIExitCodes
-from node_cli.utils.texts import Texts
+from node_cli.utils.texts import safe_load_texts
 
 logger = logging.getLogger(__name__)
-TEXTS = Texts()
+TEXTS = safe_load_texts()
 BLUEPRINT_NAME = 'node'
