+++ node_cli/main.py
@@ -43,7 +43,8 @@ from node_cli.cli.mirage_node import mirage_node_cli
 from node_cli.core.host import init_logs_dir
 from node_cli.utils.node_type import NodeType
 from node_cli.configs import LONG_LINE
-from node_cli.utils.helper import safe_load_texts, init_default_logger
+from node_cli.utils.helper import init_default_logger
+from node_cli.utils.texts import safe_load_texts
 from node_cli.utils.helper import error_exit
 
 TEXTS = safe_load_texts()
