+++ node_cli/utils/texts.py
@@ -1,6 +1,16 @@
 #   -*- coding: utf-8 -*-
 #
 #   This file is part of node-cli
+#
+#   Copyright (C) 2025-Present SKALE Labs
+#
+#   This program is free software: you can redistribute it and/or modify
+#   it under the terms of the GNU Affero General Public License as published by
+#   the Free Software Foundation, either version 3 of the License, or
+#   (at your option) any later version.
+#
+#   This program is distributed in the hope that it will be useful,
+#   but WITHOUT ANY WARRANTY; without even the implied warranty of
 #   MERCHANTABILITY or FITNESS FOR A PARTICULAR PURPOSE.  See the
 #   GNU Affero General Public License for more details.
 #
@@ -11,16 +21,9 @@ import yaml
 from node_cli.configs import TEXT_FILE
 
 
-class Texts:
-    def __init__(self):
-        self._texts = self._load()
-
-    def __getitem__(self, key):
-        return self._texts.get(key)
-
-    def _load(self):
-        with open(TEXT_FILE, 'r') as stream:
-            try:
-                return yaml.safe_load(stream)
-            except yaml.YAMLError as exc:
-                print(exc)
+def safe_load_texts():
+    with open(TEXT_FILE, 'r') as stream:
+        try:
+            return yaml.safe_load(stream)
+        except yaml.YAMLError as exc:
+            print(exc)
